refactor: split sqlite.go into multiple domain-specific files
Divide the sqlite.go file into several smaller files to improve code readability
and maintainability. Functionality is organized by domain as follows:

- sqlite_repo.go: Base structure and initialization
- sqlite_habits.go: Methods for habits
- sqlite_habit_logs.go: Methods for habit logs
- sqlite_mood.go: Methods for mood entries
- sqlite_caffeine_beverages.go: Methods for caffeine beverages
- sqlite_caffeine_intake.go: Methods for caffeine intake
- sqlite_stats.go: Methods for statistics

This refactoring doesn't change any functionality, it just reorganizes the code
to make it easier to understand and maintain.
